--- compiler-plugin-lib/src/main/kotlin/com/stslex/compiler_plugin_lib/utils/CompilerConsts.kt
@@ -1,6 +1,0 @@
-package com.stslex.compiler_plugin_lib.utils
-
-internal object CompilerConsts {
-
-    const val PATH: String = "com.stslex.compiler_plugin_lib"
-}
