--- compiler-plugin-lib/src/main/resources/META-INF/gradle-plugins/com.stslex.compiler-plugin.properties
@@ -1,1 +1,0 @@
-implementation-class=com.stslex.compiler_plugin_lib.WorkshopCompilerPlugin
