--- compiler-plugin-lib/src/main/resources/META-INF/services/org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
@@ -1,1 +1,0 @@
-com.stslex.compiler_plugin_lib.DistinctChangeIrGenerationExtension
