--- compiler-plugin-lib/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
@@ -1,1 +1,0 @@
-com.stslex.compiler_plugin_lib.WorkshopCompilerPlugin
