+++ compiler-plugin/src/main/kotlin/io/github/stslex/compiler_plugin/utils/CompilerConsts.kt
@@ -1,0 +1,6 @@
+package io.github.stslex.compiler_plugin.utils
+
+internal object CompilerConsts {
+
+    const val PATH: String = "io.github.stslex.compiler_plugin"
+}
