+++ compiler-plugin/src/main/resources/META-INF/gradle-plugins/com.stslex.compiler-plugin.properties
@@ -1,0 +1,1 @@
+implementation-class=io.github.stslex.compiler_plugin.WorkshopCompilerPlugin
