+++ compiler-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
@@ -1,0 +1,1 @@
+io.github.stslex.compiler_plugin.DistinctChangeIrGenerationExtension
