+++ compiler-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
@@ -1,0 +1,1 @@
+io.github.stslex.compiler_plugin.WorkshopCompilerPlugin
